Changed autobuilding
The autobuild keybind was changed in order to always run the makefile
that happens to be in the root directory of the git repo. It's a bit
hacky, but it works. This was ran in conjunction with `git config
--global alias.exec '!exec '`. This works because aliases are always run
in the top-level directory of the repo, allowing me to do `!git exec
make` and have it use the makefile at the root regardless of where I am
in the tree. This is useful for projects that happen to contain files in
directories not at the root of the repo.

Signed-off-by: ZweiSpeedruns <zcm727@gmail.com>